requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
ddt==1.2.1
Deprecated==1.2.6
git+git://github.com/Rapptz/discord.py.git@18fe203#egg=discord.py
gitdb==0.6.4
gitdb2==2.0.5
GitPython==2.1.12
idna==2.8
multidict==4.5.2
PyGithub==1.43.8
PyJWT==1.7.1
pyparsing==2.4.1.1
python-dateutil==2.8.0
PyYAML==5.1.1
requests==2.22.0
six==1.12.0
smmap==0.9.0
smmap2==2.0.5
urllib3==1.25.3
websockets==6.0
wrapt==1.11.2
yarl==1.3.0
